+++ integrationTests/Commands/VirtualMachinePowerManagementCommandIntegrationTest.py
@@ -6,7 +6,7 @@ from pycommon import SynchronousTaskWaiter
 from pycommon.logging_service import LoggingService
 from pycommon.pyVmomiService import pyVmomiService
 from tests.testCredentials import TestCredentials
-from vCenterShell.commands.VirtualMachinePowerManagementCommand import VirtualMachinePowerManagementCommand
+from vCenterShell.commands.power_manager_vm import VirtualMachinePowerManagementCommand
 
 
 class VirtualMachinePowerManagementCommandIntegrationTest(TestCase):
