+++ tests/TestdvPortGroupCreator.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 from mock import Mock
 
-from vCenterShell.commands.DvPortGroupCreator import DvPortGroupCreator
+from vCenterShell.network.dvswitch.creator import DvPortGroupCreator
 
 
 class TestDvPortGroupCreator(TestCase):
