+++ tests/test_bootstrapper.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from pycommon.logging_service import LoggingService
-from vCenterShell import Bootstrapper
+from vCenterShell import bootstrap
 
 
 class TestBootstrapper(TestCase):
@@ -9,7 +9,7 @@ class TestBootstrapper(TestCase):
 
     def test_get_command_executer_service(self):
         # Arrange
-        bootstrapper = Bootstrapper.Bootstrapper()
+        bootstrapper = bootstrap.Bootstrapper()
 
         # Act
         command_executer_service = bootstrapper.get_command_executer_service()
