+++ vCenterShell/commands/get_vnic_from_vm.py
@@ -2,25 +2,22 @@
 
 import qualipy.scripts.cloudshell_scripts_helpers as helpers
 from pyVmomi import vim
-
-from vCenterShell.commands.BaseCommand import BaseCommand
 from models.VirtualNicModel import VirtualNicModel
-
 from pycommon.logger import getLogger
-_logger = getLogger(__name__)
 
+_logger = getLogger("vCenterShell")
 
 #@todo very much trivial implementation. Should be moved & expanded
 class ConnectionException(Exception):
     pass
 
 
-class NetworkAdaptersRetrieverCommand(BaseCommand):
+class NetworkAdaptersRetrieverCommand(object):
     def __init__(self, pv_service, cs_retriever_service, resource_connection_details_retriever):
         """
         :param pv_service: <pycommon.pyVmomiService obj>
-        :param cs_retriever_service:
-        :param resource_connection_details_retriever:
+        :param cs_retriever_service: <pycommon.CloudshellDataRetrieverService obj>
+        :param resource_connection_details_retriever: <pycommon.ResourceConnectionDetailsRetriever obj>
         :return:
         """
         self.pvService = pv_service
@@ -31,10 +28,10 @@ class NetworkAdaptersRetrieverCommand(BaseCommand):
         resource_att = helpers.get_resource_context_details()
 
         inventory_path_data = self.csRetrieverService.getVCenterInventoryPathAttributeData(resource_att)
-        vcenter_resource_name = inventory_path_data["vCenter_resource_name"]
-        vcenter_resource_path = inventory_path_data["vm_folder"]
+        resource_name = inventory_path_data["vCenter_resource_name"]
+        resource_path = inventory_path_data["vm_folder"]
 
-        connection_details = self.resourceConnectionDetailsRetriever.get_connection_details(vcenter_resource_name)
+        connection_details = self.resourceConnectionDetailsRetriever.get_connection_details(resource_name)
         message_details = u"'{}:{}' User: '{}'".format(connection_details.host,
                                                        connection_details.port,
                                                        connection_details.user)
@@ -49,8 +46,21 @@ class NetworkAdaptersRetrieverCommand(BaseCommand):
 
         _logger.debug(u"Successfully log in {}".format(message_details))
 
-        _logger.debug(u"Retrieving... Path: '{}' Name: '{}'".format(vcenter_resource_path, vcenter_resource_name))
-        vm_machine = self.pvService.find_network_by_name(si, vcenter_resource_path, vcenter_resource_name)
+        return NetworkAdaptersRetrieverCommand.retrieve(self.pvService, si, resource_path, resource_name)
+
+
+    @staticmethod
+    def retrieve(pvService, si, path, network_name):
+        """
+        Retrieve Network by Name
+        :param pv_service: <pycommon.pyVmomiService obj>
+        :param si: <service instance>
+        :param path: <str>
+        :param network_name: <str>
+        :return: <list of 'VirtualNicModel'>
+        """
+        _logger.debug(u"Retrieving Network... Path: '{}' Name: '{}'".format(path, network_name))
+        vm_machine = pvService.find_network_by_name(si, path, network_name)
 
         result = [VirtualNicModel(x.deviceInfo.summary,
                                   x.macAddress,
@@ -58,9 +68,6 @@ class NetworkAdaptersRetrieverCommand(BaseCommand):
                                   x.connectable.startConnected)
                     for x in vm_machine.config.hardware.device
                     if isinstance(x, vim.vm.device.VirtualEthernetCard)] if vm_machine else None
+        _logger.debug(u"Retrieving Network Result: {}".format(result))
 
-        _logger.debug(u"Result: {}".format(result))
         return result
-
-
-
