+++ vCenterShell/commands/power_manager_vm.py
