+++ vCenterShell/network/dvswitch/name_generator.py
